ci: run some E2E AWS CI tests with the master using TLS
In order to get CI testing of TLS support, this change enables TLS for
certain tests in the CircleCI config.

To support this, `det-deploy aws up` now exposes port 8443 on the
instance firewall (in addition to port 8080) and accepts
`--master-tls-cert` and `--master-tls-key` options that specify a TLS
cert and key to pass to the master.

The CircleCI config is updated in a number of places to allow toggling
the TLS support per test and threading that and the follow-up
information (the port and scheme) through to everywhere it's needed.

The remainder of the change is tweaking some of the Python cert handling
code to better handle unverified connections, which is necessary since
the cert is being generated and served without being distributed to the
clients beforehand.